Seperated feature files into 2 files. clientcredentials-<featurename>.feature resourceowner-<featurename>.feature
In some instances there were only scenarios for a single credential type.
I would expect that we will be adding some additional scenario definitions soon, but is not the scope of this work.

The single exception to this is authentication.feature. There is no need to delineate credentials types as seperate scopes

This is to to make credentials types more apparent as a stronger influence in the results that you can and can not get.
Updated a few scenario names to be more behavior like in wording
Updated Feature descriptions to reflect the scenarios, in some cases the wording appeared to be copy pasted from other features. It's not like I would do anything like that...or something ;)